Move getPositionBalanceUsd to dedicated file
